+++ dsp/dsp-system/src/main/java/org/eclipse/dataspacetck/dsp/system/message/OdrlConstants.java
@@ -13,7 +13,7 @@
  *
  */
 
-package org.eclipse.dataspacetck.dsp.system.api.message;
+package org.eclipse.dataspacetck.dsp.system.message;
 
 /**
  * ODRL constants.
